Grab all covers at once from the Famicase 2017 contest
Context :
Each year the famicase contest allows contestants to create original
cartdrige
covers for non existing games using famicom cart format. Following this
first contest a
second one takes place :  A Game By Its Cover. For this second one I
needed the arts.

Why this script ?
The famicase website is a bit slow and it's not easy to peruse all the
covers at once.
So I initially created this poorly written script to grab all covers
from the website into
a local folder. Then I decded to get the html files too as some artists
had an idea about
what type of game the art should illustrate and describe it in the html
file.
Unfortunately the description text was'nt displayed because of the lack
of css.
As I am lazy I just altered the batch to grab the index file and
everything it needs to
display proper data.
Hilariously doing so mixed up the so wanted covers with thumbs, artist
pp and such :sic:
So I ended up creating a new folder Famicases which only contains the
covers. Pfiou !

Oh and one more thing : this is slow as hell, go grab a coffee the 165
covers take about 7 minutes to grab.